+++ src/adapter/graphql/src/mutations/flows_mut/flows_mut_utils.rs
@@ -67,7 +67,13 @@ pub(crate) async fn ensure_expected_dataset_kind(
     ctx: &Context<'_>,
     dataset_handle: &odf::DatasetHandle,
     dataset_flow_type: DatasetFlowType,
+    flow_run_configuration_maybe: Option<&FlowRunConfiguration>,
 ) -> Result<Option<FlowIncompatibleDatasetKind>> {
+    if let Some(FlowRunConfiguration::Compaction(CompactionConditionInput::MetadataOnly(_))) =
+        flow_run_configuration_maybe
+    {
+        return Ok(None);
+    }
     let dataset_flow_type: kamu_flow_system::DatasetFlowType = dataset_flow_type.into();
     match dataset_flow_type.dataset_kind_restriction() {
         Some(expected_kind) => {
@@ -170,16 +176,6 @@ pub(crate) async fn ensure_flow_preconditions(
         }
     }
     Ok(None)
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-pub(crate) fn ensure_set_config_flow_supported(
-    dataset_flow_type: DatasetFlowType,
-    flow_configuration_type: &'static str,
-) -> bool {
-    let dataset_flow_type: kamu_flow_system::DatasetFlowType = dataset_flow_type.into();
-    dataset_flow_type.config_restriction(flow_configuration_type)
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////